not so minor fixes
